+++ blocks/src/_common/style.scss
@@ -1,10 +1,23 @@
 /** Code Blocks **/
 
+// Featured Images
 body.post-type-post_type_characters .wp-block-post-featured-image,
 body.post-type-post_type_actors .wp-block-post-featured-image {
 	float: left;
 	width: 350px;
 	margin-right: 2rem;
+}
+
+// Undefined wrapper.
+.editor-styles-wrapper .undefined {
+	margin-left: auto;
+	margin-right: auto;
+}
+
+html :where(.undefined) {
+	margin-bottom: 28px;
+	margin-top: 28px;
+	max-width: 840px;
 }
 
 /**
