+++ blocks/src/affiliate-grid/block.json
@@ -1,0 +1,14 @@
+{
+	"$schema": "https://json.schemastore.org/block.json",
+	"apiVersion": 2,
+	"name": "lwtv/affiliate-grid",
+	"version": "1.0",
+	"title": "Affiliate Grid",
+	"category": "lezwatch",
+	"description": "A block for showing a grid of all affiliates.",
+	"keywords": [ "affiliates" ],
+	"editorScript": "file:./block.js",
+	"supports": {
+		"customClassName": true
+	}
+}
