--- blocks/src/affiliate-grid/index.js
@@ -1,2 +1,0 @@
-import './affiliate-grid.js';
-import './affiliate-item.js';
