+++ blocks/src/affiliate-item/block.json
@@ -1,0 +1,41 @@
+{
+	"$schema": "https://json.schemastore.org/block.json",
+	"apiVersion": 2,
+	"version": "1.0.0",
+	"name": "lwtv/affiliate-item",
+	"title": "Affiliate Grid Item",
+	"parent": [ "lwtv/affiliate-grid" ],
+	"category": "layout",
+	"icon": "editor-rtl",
+	"description": "This is a general Affiliate item",
+	"keywords": [
+		"affiliate",
+		"affiliate item",
+		"affiliate content"
+	],
+	
+	"supports": {
+		"customClassName": true,
+		"html": false,
+		"anchor": true
+	},
+	"attributes": {
+		"name": {
+			"type": "string",
+			"default": "Example Affiliate Name"
+		},
+		"url": {
+			"type": "string",
+			"default": "https://example.com"
+		},
+		"descr": {
+			"type": "string",
+			"default": "We are cool! Shop here!"
+		},
+		"imgUrl": {
+			"type": "string",
+			"default": "/wp-content/plugins/lwtv-plugin/assets/images/affiliate-grid.png"
+		}
+	},
+	"editorScript": "file:block.js"
+}
